MVP finalizada
falta fazer algumas funções plus para o sistema

+++ src/main/java/br/com/fiap/gs1/repositories/ViagemRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import br.com.fiap.gs1.models.Viagem;
 
 @Repository
-public interface ViagemRepository extends JpaRepository<Viagem, Integer> {
+public interface ViagemRepository extends JpaRepository<Viagem, Long> {
 
 }
